refactor: Moved datetime functions into a date API and simplified driver code
