+++ firmware/src/elevator.ccspjt
@@ -40,7 +40,7 @@ enabled=0
 chip=1
 D1=
 V1=
-buildcount=357
+buildcount=359
 
 [main.c]
 Type=4
@@ -59,7 +59,7 @@ Fixed=
 [Opened Files]
 1=main.c
 t1=1
-Count=16
+Count=19
 2=drivers\adc.h
 t2=1
 3=system.h
@@ -90,6 +90,12 @@ t14=1
 t15=1
 16=sched\scheduler.h
 t16=1
+17=api\cabin_display.c
+t17=61
+18=api\date.c
+t18=61
+19=api\date.h
+t19=3
 [..\output\misc_IOs.c]
 Type=4
 
@@ -309,7 +315,7 @@ Type=4
 [..\..\..\UART_int_Tasks\Firmware\output\elevator.c]
 Type=4
 [Units]
-Count=14
+Count=15
 1=main
 2=hardware
 3=system
@@ -324,4 +330,5 @@ Count=14
 12=api\elevator
 13=api\buttons
 14=api\queue
+15=api\date
 Link=1
